[DOCUMENTATION] documentation for HallOfFame classe

+++ src/VueCpp/GuiHallOfFame.cpp
@@ -2,44 +2,47 @@
 
 GuiHallOfFame::GuiHallOfFame(int WIDTH, int HEIGHT, int mouseX, int mouseY, bool clicked, bool running, SDL_Event /*evenement*/) :
     WIDTH(WIDTH), HEIGHT(HEIGHT), mouseX(mouseX), mouseY(mouseY), clicked(clicked), running(running) {
-        hofController.addObserver(this);
+        hofController.addObserver(this); // Add observer to the HofController
     }
 
 GuiHallOfFame::~GuiHallOfFame() {
-    hofController.removeObserver(this);
+    hofController.removeObserver(this); // Remove observer from the HofController
 }
 
 bool GuiHallOfFame::drawHallOfFame(int WIDTH, int HEIGHT, int mouseX, int mouseY, bool clicked, bool running, SDL_Event evenement) {
 
-    Draw draw(clicked);
-    hallOfFameWindow = draw.createWindow(WIDTH, HEIGHT, "Hall of Fame", hallOfFameWindow);
-    hallOfFameRenderer = draw.createRenderer(hallOfFameWindow, hallOfFameRenderer);
-    hofController.readTxtFile();
+    Draw draw(clicked); // Initialize Draw object with click status
+    hallOfFameWindow = draw.createWindow(WIDTH, HEIGHT, "Hall of Fame", hallOfFameWindow); // Create SDL window
+    hallOfFameRenderer = draw.createRenderer(hallOfFameWindow, hallOfFameRenderer); // Create SDL renderer
+    hofController.readTxtFile(); // Read data from file
 
     // Game loop
     while (!running) {
         // Event handling
         while (SDL_PollEvent(&evenement) != 0) {
             if (evenement.type == SDL_QUIT) {
-                running = true;
+                running = true; // Quit loop if quit event is triggered
                 return running;
             } else if (evenement.type == SDL_MOUSEBUTTONDOWN) {
-                clicked = true;
-                SDL_GetMouseState(&mouseX, &mouseY);
+                clicked = true; // Set clicked flag to true when mouse button is pressed
+                SDL_GetMouseState(&mouseX, &mouseY); // Get mouse position
             } else if (evenement.type == SDL_MOUSEBUTTONUP) {
-                clicked = false;
+                clicked = false; // Set clicked flag to false when mouse button is released
             }
-            SDL_RenderClear(hallOfFameRenderer);
+            SDL_RenderClear(hallOfFameRenderer); // Clear the renderer
 
+            // Draw background, title, and podium
             draw.drawImage(hallOfFameRenderer, 0, 0, WIDTH, HEIGHT, "assets/images/hofBg.png");
             draw.drawText(hallOfFameRenderer, 500, 60, "Hall of Fame", 50, 50, 50, 255, 40);
             draw.drawImage(hallOfFameRenderer, 200, 300, 800, 250, "assets/images/podium.png");
 
+            // Determine selected difficulty level
             difficultyHof = returnDifficulty(draw, hallOfFameRenderer, mouseX, mouseY, clicked);
 
+            // Draw Hall of Fame data
             drawData(difficultyHof, draw, hallOfFameRenderer, hofController);
 
-            SDL_RenderPresent(hallOfFameRenderer);
+            SDL_RenderPresent(hallOfFameRenderer); // Update the renderer
         }
     }
     // Destroy the renderer
@@ -51,6 +54,7 @@ bool GuiHallOfFame::drawHallOfFame(int WIDTH, int HEIGHT, int mouseX, int mouseY
     return running;
 }
 
+// Determine selected difficulty level
 int GuiHallOfFame::returnDifficulty(Draw draw, SDL_Renderer* hallOfFameRenderer, int mouseX, int mouseY, bool clicked) {
     vector<string> links;
 
@@ -62,19 +66,21 @@ int GuiHallOfFame::returnDifficulty(Draw draw, SDL_Renderer* hallOfFameRenderer,
         ss << "./assets/images/" << i << ".png";
         links.push_back(ss.str());
 
+        // Create buttons for each difficulty level
         draw.createButton(hallOfFameRenderer, x, 180, 50, 50, links.back().c_str(), mouseX, mouseY, clicked, [&](){
             cout << "Difficulte choisie : " << i << endl;
-            difficultyHof = i;
+            difficultyHof = i; // Update selected difficulty level
         });
     }
     return difficultyHof;
 }
 
+// Draw Hall of Fame data
 void GuiHallOfFame::drawData(int difficultyHof, Draw draw, SDL_Renderer* hallOfFameRenderer, HofController hofController) {
-    hofController.topThreeSortedData(difficultyHof);
+    hofController.topThreeSortedData(difficultyHof); // Get top three sorted data for the selected difficulty level
 
-
-    for (long long unsigned int i = 0; i < hofController.getSortedDataFoh().size(); i++) {
+    // Display top three players' data
+    for (size_t i = 0; i < hofController.getSortedDataFoh().size(); i++) {
         stringstream ss;
         ss << hofController.getSortedDataFoh()[i].name << " - " << hofController.getSortedDataFoh()[i].level;
         if (i == 0) {
@@ -90,6 +96,7 @@ void GuiHallOfFame::drawData(int difficultyHof, Draw draw, SDL_Renderer* hallOfF
     }
 }
 
+// Update method from Observer interface
 void GuiHallOfFame::update() {
-    hofController.readTxtFile();
+    hofController.readTxtFile(); // Read updated data from file
 }
